Refactor license management and email verification hooks
- Updated SendVerificationHook to use new data structures for User and VerificationTargetSet.
- Enhanced LicenseCheckHook to utilize jOOQ for database interactions, replacing JPA EntityManager.
- Refactored LicenseDeleteHook to delete user licenses and associated activities using jOOQ.
- Modified LicenseUpdateHook to create and update user licenses and activities with jOOQ.
- Converted LicenseLastActivity and UserLicense entities to records for better immutability and simplicity.
- Added .env.development file for UI configuration with JWT secret and API server URL.
- Updated package.json to include jsonwebtoken types and library.
- Introduced vite-plugin-bouncr-auth for local development, handling JWT and proxying requests.
- Adjusted vite.config.ts to integrate the new authentication plugin.
- Updated pom.xml to reflect dependency changes and version updates, including migration to jOOQ.

+++ bouncr-components/src/main/java/db/migration/V21__CreateUserProfiles.java
@@ -35,7 +35,7 @@ public class V21__CreateUserProfiles extends BaseJavaMigration {
                     .column(field("min_length", SQLDataType.SMALLINT.nullable(true)))
                     .column(field("max_length", SQLDataType.SMALLINT.nullable(true)))
                     .column(field("needs_verification", SQLDataType.BOOLEAN.nullable(false)))
-                    .column(field("position", SQLDataType.TINYINT.nullable(false)))
+                    .column(field("position", SQLDataType.SMALLINT.nullable(false)))
                     .constraints(
                             constraint().primaryKey(field("user_profile_field_id")),
                             constraint().unique(field("name")),
